Generic ServiceClient
